# Benchmark_store_SelectSample: stop accumulating results