Appending to the same slice over all sub-benchmarks means you get different overhead from the first benchmark to the last as the slice gets bigger.

The accumulated slice was only used to print a log line, which we can do more neatly using `ReportMetric()`, and without creating nearly a gigabyte of garbage.

Before:
```
Benchmark_store_SelectSample/count_over_time({foo="bar"}[5m])-4                        1        5640722394 ns/op        2531730112 B/op   213464 allocs/op
Benchmark_store_SelectSample/rate({foo="bar"}[5m])-4                                   1        5719734486 ns/op        2489809096 B/op   211765 allocs/op
Benchmark_store_SelectSample/bytes_rate({foo="bar"}[5m])-4                             1        5776751945 ns/op        2609764312 B/op   212025 allocs/op
Benchmark_store_SelectSample/bytes_over_time({foo="bar"}[5m])-4                        1        6617705175 ns/op        2331137296 B/op   212763 allocs/op
2023/01/23 14:34:05 sample processed 25259920
```

After:
```
Benchmark_store_SelectSample/count_over_time({foo="bar"}[5m])-4                        1        5216947643 ns/op           6314980 samples/op   1709845624 B/op   214137 allocs/op
Benchmark_store_SelectSample/rate({foo="bar"}[5m])-4                                   1        5328391412 ns/op           6314980 samples/op   1706996944 B/op   213205 allocs/op
Benchmark_store_SelectSample/bytes_rate({foo="bar"}[5m])-4                             1        5291661130 ns/op           6314980 samples/op   1708740728 B/op   212596 allocs/op
Benchmark_store_SelectSample/bytes_over_time({foo="bar"}[5m])-4                        1        5251616536 ns/op           6314980 samples/op   1709636168 B/op   212240 allocs/op
```

There are another nearby benchmarks `Benchmark_store_SelectLogsRegexBackward` also accumulating results and printing log messages, but they start from scratch each time.

**Checklist**
- [x] Reviewed the [`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md) guide (**required**)
- NA Documentation added
- [x] Tests updated
- NA `CHANGELOG.md` updated
- NA Changes that require user attention or interaction to upgrade are documented in `docs/sources/upgrading/_index.md`